Added mapview component to homepage + moving mapview.js to components folder
